'Stable' x2
Added some authorization for admin and done the path of cart service

+++ SS.Template.Api/Controllers/ProductsController.cs
@@ -60,34 +60,50 @@ namespace SS.Template.Api.Controllers
 
         // POST: api/products
         //Admin Route
+        [Authorize]
         [HttpPost]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Save([FromBody] ProductsModel products)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return BadRequest("Not Authorized");
+            }
             await _productsService.Create(products);
             return Ok();
         }
 
         // PUT: api/products/5
         //Admin Route
+        [Authorize]
         [HttpPut("{id:guid}")]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Update(Guid id, [FromBody] ProductsModel products)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return BadRequest("Not Authorized");
+            }
             await _productsService.Update(id, products);
             return Ok();
         }
 
         // DELETE: api/productss/5
         //Admin Route
+        [Authorize]
         [HttpDelete("{id:guid}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return BadRequest("Not Authorized");
+            }
             await _productsService.Delete(id);
             return Ok();
         }
